Simplify mysql step of dev deploy and add docs for copying images
The docs suggested passing mysql options as arguments and the values had to be
extracted from the config file. A cleaner way is to just pass the config
file instead. We also did not have the docs to explain how to copy images
into the Artifact Registry, which has now been added.